+++ tomcat/src/main/java/org/apache/NotFoundController.java
@@ -1,8 +1,8 @@
-package com.techcourse.controller;
+package org.apache;
 
+import com.techcourse.controller.AbstractRequestController;
 import com.techcourse.exception.UncheckedServletException;
 import java.nio.charset.StandardCharsets;
-import org.apache.ResourceReader;
 import org.apache.coyote.HttpRequest;
 import org.apache.coyote.HttpResponse;
