+++ tomcat/src/test/java/com/techcourse/controller/NotFoundControllerTest.java
@@ -6,6 +6,7 @@ import java.io.File;
 import java.net.URL;
 import java.nio.file.Files;
 import java.util.List;
+import org.apache.NotFoundController;
 import org.apache.coyote.HttpRequest;
 import org.apache.coyote.HttpResponse;
 import org.apache.coyote.RequestHandler;
